Refactor: Moved Files to web folder
