+++ src/web/components/Background.styles.js
