+++ src/web/components/Banner.test.js
