+++ src/web/view/home/Rules.js
