+++ src/web/view/home/Rules.test.js
