+++ src/web/view/play/Play.test.js
@@ -4,7 +4,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import { Provider } from 'react-redux';
 import { Play } from './Play';
 
-import { rootReducer } from '../../store/rootReducer';
+import { rootReducer } from '../../../store/rootReducer';
 
 const store = configureStore({ reducer: rootReducer });
